requirements.txt
flask-redis
watchdog
Pillow